requirements.txt
opencv-python
numpy
matplotlib
easyocr
ultralytics
mlem
fastapi
uvicorn
streamlit
streamlit_pydantic
pillow